features: export uprobe/kprobe link features
The use of multi-attach for uprobes/kprobes requires both loading the
program with the appropriate attach type, and then using the correct
call to do the attach. Sometimes the program loading is not tightly
coupled to the attach so it's painful to have to fail later and then
re-load the program. More generally, forcing consumers of this package
to track their own feature availability on top of which things return
unsupported errors is bad -- we should make most feature probing
available and exported to the extent possible.

This PR exports from `features`:
 * `HaveBPFLinkUprobeMulti`
 * `HaveBPFLinkKprobeMulti`
 * `HaveBPFLinkKprobeSession`
