+++ link/uprobe_multi.go
@@ -8,8 +8,7 @@ import (
 	"os"
 
 	"github.com/cilium/ebpf"
-	"github.com/cilium/ebpf/asm"
-	"github.com/cilium/ebpf/internal"
+	"github.com/cilium/ebpf/features"
 	"github.com/cilium/ebpf/internal/sys"
 	"github.com/cilium/ebpf/internal/unix"
 )
@@ -108,7 +107,7 @@ func (ex *Executable) uprobeMulti(symbols []string, prog *ebpf.Program, opts *Up
 	}
 
 	if err != nil {
-		if haveFeatErr := haveBPFLinkUprobeMulti(); haveFeatErr != nil {
+		if haveFeatErr := features.HaveBPFLinkUprobeMulti(); haveFeatErr != nil {
 			return nil, haveFeatErr
 		}
 		return nil, err
@@ -175,46 +174,3 @@ var _ Link = (*uprobeMultiLink)(nil)
 func (kml *uprobeMultiLink) Update(_ *ebpf.Program) error {
 	return fmt.Errorf("update uprobe_multi: %w", ErrNotSupported)
 }
-
-var haveBPFLinkUprobeMulti = internal.NewFeatureTest("bpf_link_uprobe_multi", func() error {
-	prog, err := ebpf.NewProgram(&ebpf.ProgramSpec{
-		Name: "probe_upm_link",
-		Type: ebpf.Kprobe,
-		Instructions: asm.Instructions{
-			asm.Mov.Imm(asm.R0, 0),
-			asm.Return(),
-		},
-		AttachType: ebpf.AttachTraceUprobeMulti,
-		License:    "MIT",
-	})
-	if errors.Is(err, unix.E2BIG) {
-		// Kernel doesn't support AttachType field.
-		return internal.ErrNotSupported
-	}
-	if err != nil {
-		return err
-	}
-	defer prog.Close()
-
-	// We try to create uprobe multi link on '/' path which results in
-	// error with -EBADF in case uprobe multi link is supported.
-	fd, err := sys.LinkCreateUprobeMulti(&sys.LinkCreateUprobeMultiAttr{
-		ProgFd:     uint32(prog.FD()),
-		AttachType: sys.BPF_TRACE_UPROBE_MULTI,
-		Path:       sys.NewStringPointer("/"),
-		Offsets:    sys.SlicePointer([]uint64{0}),
-		Count:      1,
-	})
-	switch {
-	case errors.Is(err, unix.EBADF):
-		return nil
-	case errors.Is(err, unix.EINVAL):
-		return internal.ErrNotSupported
-	case err != nil:
-		return err
-	}
-
-	// should not happen
-	fd.Close()
-	return errors.New("successfully attached uprobe_multi to /, kernel bug?")
-}, "6.6")
